moving on with purging the need for forward declarations in the game-state folder
