+++ include/game-state/game_action.h
@@ -6,6 +6,30 @@
 #include "condition.h"
 #include "item.h"
 
+// ACTION STRUCTURE DEFINITION + BASIC FUNCTIONS ------------------------------
+typedef struct game_action_effect{
+    item_t *item;
+    attribute_t* attribute_to_modify;
+    attribute_value_t new_value;
+    struct game_action_effect *next; //mandatory for utlist macros
+} game_action_effect_t;
+
+/* This typedef is to distinguish between game_action_effect_t
+* pointers which are used to point to the game_action_effect_t structs
+* in the traditional sense, and those which are used to enable UTLIST functionality
+* on the game_action_effect_t structs as specified in src/common/include
+*/
+typedef struct game_action_effect action_effect_list_t;
+
+
+typedef struct game_action {
+    UT_hash_handle hh;
+    char* action_name;
+    condition_list_t *conditions; //must be initialized to NULL
+    action_effect_list_t *effects; //must be initialized to NULL
+    char* success_str;
+    char* fail_str;
+} game_action_t;
 
 // ------------------------- ACTION FUNCTIONS -------------------------
 // the following functions are to do with searching for allowed actions
