+++ include/game-state/player.h
@@ -10,12 +10,8 @@
 #include "stats.h"
 #include "item.h"
 #include "playerclass/class.h"
-#include "skilltrees/skilltrees_enums.h"
+#include "skilltrees/skilltrees_common.h"
 #include "skilltrees/inventory.h"
-#include "battle/battle_structs.h"
-
-/* Forward declaration for skilltrees */
-typedef struct skill skill_t;
 
 /* A player in game */
 typedef struct player {
